home-automation/planner-interface/requirements.txt
requests==2.32.3
google-auth==2.34.0
google-auth-oauthlib==1.2.1